fix(#99 WI-3): Gate-4 r1 — provider-fetch lifetime (clear-before-fetch + dismiss guard); funnel observes the new notification; popover rows + pill style split (file budget); doc drift
